Move stats utils to `p2p` module
