+++ p2p/stats/stddev.py
